rename QuickActions
